refactor: common 모듈로 dto클래스 이동
