+++ domain/common/src/main/java/com/kth/message/dto/weather/request/ActionDto.java
@@ -1,4 +1,4 @@
-package com.kth.message.weather.web.dto.request;
+package com.kth.message.dto.weather.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
