+++ domain/common/src/main/java/com/kth/message/dto/weather/request/RequestActionDto.java
@@ -1,8 +1,7 @@
-package com.kth.message.weather.web.dto.request;
+package com.kth.message.dto.weather.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.kth.message.weather.web.dto.request.skill.WeatherDto;
 import lombok.Builder;
 import lombok.Getter;
