+++ domain/common/src/main/java/com/kth/message/dto/weather/request/WeatherDto.java
@@ -1,8 +1,7 @@
-package com.kth.message.weather.web.dto.request.skill;
+package com.kth.message.dto.weather.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.kth.message.weather.web.dto.request.ParamDto;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
